pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "Task Manager based on Django"
authors = ["Sergey R <goodstop687@gmail.com>"]
readme = "README.md"
packages = [{include="task_manager"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.7"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
python-dotenv = "^1.0.1"
whitenoise = {extras = ["brotli"], version = "^6.7.0"}
gunicorn = "^22.0.0"
django-bootstrap5 = "^24.2"
python-gettext = "^5.0"
django-filter = "^24.3"
rollbar = "0.16.3"
black = "^24.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.0"
coverage = "^7.6.1"